+++ tests/integration.rs
@@ -246,7 +246,9 @@ fn communication_self_test_can_fail_in_step1() {
         vec![0x56],
     )];
     let mut sensor = new_1018(&transactions);
-    sensor.communication_self_test().expect_err("Should return error");
+    sensor
+        .communication_self_test()
+        .expect_err("Should return error");
     destroy(sensor);
 }
 
@@ -258,7 +260,9 @@ fn communication_self_test_can_fail_in_step2() {
         I2cTrans::write_read(DEV_ADDR, vec![Register::DCST_RESP], vec![0xAB]),
     ];
     let mut sensor = new_1018(&transactions);
-    sensor.communication_self_test().expect_err("Should return error");
+    sensor
+        .communication_self_test()
+        .expect_err("Should return error");
     destroy(sensor);
 }
 
@@ -271,7 +275,9 @@ fn communication_self_test_can_fail_in_step3() {
         I2cTrans::write_read(DEV_ADDR, vec![Register::CTRL2], vec![BitFlags::DCST]),
     ];
     let mut sensor = new_1018(&transactions);
-    sensor.communication_self_test().expect_err("Should return error");
+    sensor
+        .communication_self_test()
+        .expect_err("Should return error");
     destroy(sensor);
 }
 
@@ -285,6 +291,24 @@ fn communication_self_test_can_fail_in_step4() {
         I2cTrans::write_read(DEV_ADDR, vec![Register::DCST_RESP], vec![0x56]),
     ];
     let mut sensor = new_1018(&transactions);
-    sensor.communication_self_test().expect_err("Should return error");
+    sensor
+        .communication_self_test()
+        .expect_err("Should return error");
+    destroy(sensor);
+}
+
+#[test]
+fn can_enable_mems_self_test() {
+    let transactions = [I2cTrans::write(DEV_ADDR, vec![Register::SELF_TEST, 0xCA])];
+    let mut sensor = new_1018(&transactions);
+    sensor.enable_mems_self_test().unwrap();
+    destroy(sensor);
+}
+
+#[test]
+fn can_disable_mems_self_test() {
+    let transactions = [I2cTrans::write(DEV_ADDR, vec![Register::SELF_TEST, 0])];
+    let mut sensor = new_1018(&transactions);
+    sensor.disable_mems_self_test().unwrap();
     destroy(sensor);
 }