+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Added
+- Enable/disable MEMS self-test function.
+
 ### Changed
 - [breaking-change] Renamed communication self-test function `self_test()` to
 `communication_self_test()`.